backend/requirements.txt
fastapi
uvicorn
pandas
numpy
yfinance
fundamentus
statsmodels
scipy
SQLAlchemy
psycopg2-binary
python-dotenv

